requirements.txt
lirc==2.0.1
paho-mqtt==1.6.1
